Convert all entries on sys.path to absolute paths, and also update the __file__ attributes of already-imported modules to be absolute.  This helps robustify the interpreter against os.chdir() calls from the application.
Only remove setdefaultencoding() from sys if it exists; if this module is
run as a script (since there is a _test() function that gets run), it broke
because the script attempts to remove it again after the import of site
has already done so.  This allows the module to be run as a script again.

makepath():  New function, standardizes all pathname normalization in one
             place.

--HG--
branch : legacy-trunk