[ Bug #116636 ] Bug in StringIO.write() http://sourceforge.net/bugs/?func=detailbug&bug_id=116636&group_id=5470 bobalex@rsv.ricoh.com
Bug report: If the file position is less than the end of the "file",
and a write is performed extending past then end of the file, the data
string is corrupted.

Solution: in write(), when writing past the end, properly set self.len
when newpos is > self.len.

--HG--
branch : legacy-trunk